[t] the end-to-end tests for main() and read_options() were crap
The tests weren't actually testing the functionality which is desired.  We
want that the program starts and generates a .tex file.  We also want that
the command line options are accepted and generate some kind of appropriate
output (which can then be tested).  *This* is the kind of stuff we need here.